+++ meteor/client/mozdef.js
@@ -31,6 +31,7 @@ if (Meteor.isClient) {
         Session.set('attackerlimit','10');
         Session.set('attackersearchIP','');
         Session.set('blockIPipaddress','');
+        Session.set('watchItemwatchcontent','');
         Session.set('blockFQDN','');
         getAllPlugins();
 
@@ -331,13 +332,18 @@ if (Meteor.isClient) {
             dshielditem=$("<li><a class='ipmenu-dshield' data-ipaddress='" + iptext + "'href='#'>dshield</a></li>");
             intelitem=$("<li><a class='ipmenu-intel' data-ipaddress='" + iptext + "'href='#'>ip intel</a></li>");
             searchitem=$("<li><a class='ipmenu-search' data-ipaddress='" + iptext + "'href='#'>search kibana</a></li>");
+            if ( isFeature('watchItem') ){
+            watchItemitem=$("<li><a class='ipmenu-watchitem' data-ipaddress='" + iptext + "'href='#'>watch</a></li>");
+            }else{
+            watchItemitem=$();
+            }
             if ( isFeature('blockip') ){
                 blockIPitem=$("<li><a class='ipmenu-blockip' data-ipaddress='" + iptext + "'href='#'>block</a></li>");
             }else{
                 blockIPitem=$();
             }
+            ipmenu.append(copyitem,whoisitem,dshielditem,intelitem,searchitem,watchItemitem,blockIPitem);
 
-            ipmenu.append(copyitem,whoisitem,dshielditem,intelitem,searchitem,blockIPitem);
 
             $(this).parent().parent().append(ipmenu);
         });