+++ mq/esworker_sns_sqs.py
@@ -47,15 +47,8 @@ class taskConsumer(object):
         self.connection = mqConnection
         self.esConnection = esConnection
         self.taskQueue = taskQueue
-
         self.pluginList = registerPlugins()
-
         self.options = options
-
-        if self.options.esbulksize != 0:
-            # if we are bulk posting enable a timer to occasionally flush the bulker even if it's not full
-            # to prevent events from sticking around an idle worker
-            self.esConnection.start_bulk_timer()
 
     def run(self):
         self.taskQueue.set_message_class(RawMessage)