+++ mq/plugins/complianceitems.py
@@ -18,16 +18,16 @@ class message(object):
         """
         for key in ['target', 'policy', 'check', 'compliance',
                     'link', 'utctimestamp']:
-            if key not in item.keys():
+            if key not in item:
                 return False
         for key in ['level', 'name', 'url']:
-            if key not in item['policy'].keys():
+            if key not in item['policy']:
                 return False
         for key in ['description', 'location', 'name', 'test']:
-            if key not in item['check'].keys():
+            if key not in item['check']:
                 return False
         for key in ['type', 'value']:
-            if key not in item['check']['test'].keys():
+            if key not in item['check']['test']:
                 return False
         return True