+++ tests/mozdef_util/test_bulk_queue.py
@@ -15,7 +15,7 @@ class BulkQueueTest(UnitTestSuite):
         super(BulkQueueTest, self).setup()
 
     def num_objects_saved(self):
-        self.flush(self.event_index_name)
+        self.refresh(self.event_index_name)
         search_query = SearchQuery()
         search_query.add_must(ExistsMatch('keyname'))
         results = search_query.execute(self.es_client)
@@ -96,18 +96,18 @@ class TestTimer(BulkQueueTest):
     def test_basic_timer(self):
         queue = BulkQueue(self.es_client, flush_time=2)
         assert queue.started() is False
-        queue.start_timer()
+        queue.start_thread()
         assert queue.started() is True
         queue.add(index='events', doc_type='_doc', body={'keyname': 'valuename'})
         assert queue.size() == 1
         time.sleep(3)
         assert queue.size() == 0
-        queue.stop_timer()
+        queue.stop_thread()
         assert queue.started() is False
 
     def test_over_threshold(self):
         queue = BulkQueue(self.es_client, flush_time=3, threshold=10)
-        queue.start_timer()
+        queue.start_thread()
         for num in range(0, 201):
             queue.add(index='events', doc_type='_doc', body={'keyname': 'value' + str(num)})
         assert self.num_objects_saved() == 200
@@ -115,11 +115,11 @@ class TestTimer(BulkQueueTest):
         time.sleep(4)
         assert self.num_objects_saved() == 201
         assert queue.size() == 0
-        queue.stop_timer()
+        queue.stop_thread()
 
     def test_two_iterations(self):
         queue = BulkQueue(self.es_client, flush_time=3, threshold=10)
-        queue.start_timer()
+        queue.start_thread()
         for num in range(0, 201):
             queue.add(index='events', doc_type='_doc', body={'keyname': 'value' + str(num)})
         assert self.num_objects_saved() == 200
@@ -132,11 +132,11 @@ class TestTimer(BulkQueueTest):
         assert self.num_objects_saved() == 401
         time.sleep(3)
         assert self.num_objects_saved() == 402
-        queue.stop_timer()
+        queue.stop_thread()
 
     def test_ten_iterations(self):
         queue = BulkQueue(self.es_client, flush_time=3, threshold=10)
-        queue.start_timer()
+        queue.start_thread()
         total_events = 0
         for num_rounds in range(0, 10):
             for num in range(0, 20):
@@ -144,5 +144,5 @@ class TestTimer(BulkQueueTest):
                 queue.add(index='events', doc_type='_doc', body={'keyname': 'value' + str(num)})
             assert self.num_objects_saved() == total_events
         assert queue.size() == 0
-        queue.stop_timer()
+        queue.stop_thread()
         assert self.num_objects_saved() == 200