+++ tests/mozdef_util/test_elasticsearch_client.py
@@ -26,7 +26,7 @@ class ElasticsearchClientTest(UnitTestSuite):
         self.es_client = ElasticsearchClient(self.options.esservers, bulk_refresh_time=3)
 
     def get_num_events(self):
-        self.flush('events')
+        self.refresh('events')
         search_query = SearchQuery()
         search_query.add_must(TermMatch('type', 'event'))
         search_query.add_aggregation(Aggregation('type'))
@@ -95,8 +95,8 @@ class TestWriteWithRead(ElasticsearchClientTest):
             'url': 'https://mozilla.org',
             'utctimestamp': '2016-08-19T16:40:57.851092+00:00'
         }
-        self.saved_alert_result = self.es_client.save_alert(body=self.alert)
-        self.flush('alerts')
+        self.saved_alert = self.es_client.save_alert(body=self.alert)
+        self.refresh('alerts')
 
     def test_saved_type(self):
         assert self.saved_alert_result['_type'] == '_doc'
@@ -146,7 +146,7 @@ class TestSimpleWrites(ElasticsearchClientTest):
             self.es_client.save_event(body=event)
 
         assert mock_class.request_counts == 100
-        self.flush(self.event_index_name)
+        self.refresh(self.event_index_name)
         num_events = self.get_num_events()
         assert num_events == 100
 
@@ -154,7 +154,7 @@ class TestSimpleWrites(ElasticsearchClientTest):
         event = json.dumps({"key": "example value for string of json test"})
         self.es_client.save_event(body=event)
 
-        self.flush(self.event_index_name)
+        self.refresh(self.event_index_name)
         num_events = self.get_num_events()
         assert num_events == 1
 
@@ -172,7 +172,7 @@ class TestSimpleWrites(ElasticsearchClientTest):
         event = json.dumps({"key.othername": "example value for string of json test"})
         self.es_client.save_event(body=event)
 
-        self.flush(self.event_index_name)
+        self.refresh(self.event_index_name)
         num_events = self.get_num_events()
         assert num_events == 1
 
@@ -189,7 +189,7 @@ class TestSimpleWrites(ElasticsearchClientTest):
         query = SearchQuery()
         default_event = {}
         self.populate_test_event(default_event)
-        self.flush(self.event_index_name)
+        self.refresh(self.event_index_name)
 
         query.add_must(ExistsMatch('summary'))
         results = query.execute(self.es_client)
@@ -224,7 +224,7 @@ class TestSimpleWrites(ElasticsearchClientTest):
             }
         }
         self.populate_test_event(default_event)
-        self.flush(self.event_index_name)
+        self.refresh(self.event_index_name)
 
         query.add_must(ExistsMatch('summary'))
         results = query.execute(self.es_client)
@@ -247,7 +247,7 @@ class TestSimpleWrites(ElasticsearchClientTest):
             }
         }
         self.populate_test_event(default_event)
-        self.flush(self.event_index_name)
+        self.refresh(self.event_index_name)
 
         query.add_must(ExistsMatch('summary'))
         results = query.execute(self.es_client)
@@ -281,7 +281,7 @@ class TestBulkWrites(BulkTest):
         for event in events:
             self.es_client.save_event(body=event, bulk=True)
 
-        self.flush(self.event_index_name)
+        self.refresh(self.event_index_name)
         time.sleep(1)
 
         # We encountered a weird bug in travis
@@ -305,21 +305,21 @@ class TestBulkWritesWithMoreThanThreshold(BulkTest):
         for event in events:
             self.es_client.save_object(index='events', doc_type='event', body=event, bulk=True)
 
-        self.flush(self.event_index_name)
+        self.refresh(self.event_index_name)
 
         # We encountered a weird bug in travis
         # that would sometimes cause the number
         # of requests sent to ES to fluctuate.
         # As a result, we're checking within 5 requests
         # from 20, to verify we are still using bulk
-        non_flushed_request_count = self.mock_class.request_counts
+        non_refreshed_request_count = self.mock_class.request_counts
         assert self.mock_class.request_counts <= 25 and self.mock_class.request_counts >= 15
         assert self.get_num_events() == 1900
         time.sleep(5)
         # All we want to check here is that during the sleep
         # we purged the queue and sent the remaining events to ES
-        assert self.mock_class.request_counts > non_flushed_request_count
-        self.flush(self.event_index_name)
+        assert self.mock_class.request_counts > non_refreshed_request_count
+        self.refresh(self.event_index_name)
         assert self.get_num_events() == 1995
 
 
@@ -336,7 +336,7 @@ class TestBulkWritesWithLessThanThreshold(BulkTest):
 
         assert self.get_num_events() == 0
 
-        self.flush(self.event_index_name)
+        self.refresh(self.event_index_name)
         time.sleep(5)
         assert self.get_num_events() == 6
 
@@ -359,7 +359,7 @@ class TestWriteWithIDExists(ElasticsearchClientTest):
         event['new_key'] = 'updated_value'
         saved_event = self.es_client.save_event(body=event, doc_id=event_id)
         assert saved_event['_id'] == event_id
-        self.flush(self.event_index_name)
+        self.refresh(self.event_index_name)
         self.es_client.get_event_by_id(event_id)
 
 
@@ -501,8 +501,8 @@ class TestBulkInvalidFormatProblem(BulkTest):
             "utcstamp": "abc",
         }
 
-        self.es_client.save_event(body=event, bulk=True)
-        self.es_client.save_event(body=malformed_event, bulk=True)
-        self.flush(self.event_index_name)
+        self.es_client.save_object(index='events', doc_type='event', body=event, bulk=True)
+        self.es_client.save_object(index='events', doc_type='event', body=malformed_event, bulk=True)
+        self.refresh(self.event_index_name)
         time.sleep(5)
         assert self.get_num_events() == 1