+++ mozdef_util/HISTORY.rst
@@ -24,3 +24,14 @@ Update Geolite db location
 ------------------
 
 Add is_ip utility function
+
+
+1.0.4 (2019-01-23)
+------------------
+
+* Replaced timer with threads for cleaner bulk importing
+
+1.0.5 (2019-03-06)
+------------------
+
+* Replace elasticsearch flush with refresh